Refactored to remove common code between network_checkpoint_test.py and temporal_memory_compatibility_test.py
